# admin-api/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Unchain Admin API
  description: |
    Comprehensive OpenAPI specification for the Unchain Admin API.

    This API controls the configuration of the Unchain Feature Management platform, 
    including Features, Strategies, Projects, Environments, Context Fields, 
    Change Requests, Segments, FeatureTags, and Dependencies.

    **Base URL**: `<your-unchain-instance>/api/admin`
  version: 4.0.0
  contact:
    name: Philipp Walther
servers:
  - url: http://localhost:8080
    description: Local Instance
security:
  - OidcAuth: []
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      x-required-permissions: [PROJECT_READ]
      summary: List all projects
      tags: [Projects]
      parameters:
        - name: archived
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
    post:
      summary: Create a new project
      operationId: createProject
      x-required-permissions: [PROJECT_CREATE]
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          headers:
            location:
              description: The location of the newly created resource.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      x-required-permissions: [PROJECT_READ]
      summary: Get project details
      tags: [Projects]
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      x-required-permissions: [PROJECT_UPDATE]
      summary: Update project
      tags: [Projects]
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
    delete:
      operationId: deleteProject
      x-required-permissions: [PROJECT_DELETE]
      summary: Delete project
      tags: [Projects]
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Project deleted
...
  /segments:
    description: A segment is a reusable collection of strategy constraints. Feature flags have one or more activation strategies that define how and when a feature is shown, for example, to a percentage of users, to a specific group, or only in certain environments.
    get:
      operationId: listSegments
      x-required-permissions: [SEGMENT_READ]
      summary: List all segments
      tags: [Segments]
      responses:
        "200":
          description: List of segments
          content:
            application/json:
              schema:
                type: object
                properties:
                  segments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Segment"
    post:
      operationId: createSegment
      summary: Create a new segment
      tags: [Segments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSegmentRequest"
      responses:
        "201":
          description: Segment created
  /segments/{segmentId}:
    get:
      operationId: getSegment
      x-required-permissions: [SEGMENT_READ]
      summary: Get a segment
      tags: [Segments]
      parameters:
        - $ref: "#/components/parameters/segmentId"
      responses:
        "200":
          description: Segment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Segment"
    put:
      operationId: updateSegment
      x-required-permissions: [SEGMENT_UPDATE]
      summary: Update a segment
      tags: [Segments]
      parameters:
        - $ref: "#/components/parameters/segmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSegmentRequest"
      responses:
        "200":
          description: Segment updated
    delete:
      operationId: deleteSegment
      x-required-permissions: [SEGMENT_DELETE]
      summary: Delete a segment
      tags: [Segments]
      parameters:
        - $ref: "#/components/parameters/segmentId"
      responses:
        "200":
          description: Segment deleted

  # =========================================================================
  # CONTEXT
  # =========================================================================
...
components:
  securitySchemes:
    OidcAuth:
      type: openIdConnect
      openIdConnectUrl: https://server.redmoon.ch/auth/oidc/.well-known/openid-configuration
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
      description: The ID of the project (e.g. 'default')
...
    segmentId:
      name: segmentId
      in: path
      required: true
      schema:
        type: integer
      description: The ID of the segment
  schemas:
...
    Constraint:
      type: object
      properties:
        contextName:
          type: string
        operator:
          type: string
          enum:
            [
              IN,
              NOT_IN,
              STR_ENDS_WITH,
              STR_STARTS_WITH,
              STR_CONTAINS,
              NUM_EQ,
              NUM_GT,
              NUM_LT,
              DATE_AFTER,
              DATE_BEFORE,
              SEMVER_EQ,
              SEMVER_GT,
              SEMVER_LT,
            ]
        values:
          type: array
          items:
            type: string
        caseInsensitive:
          type: boolean
        inverted:
          type: boolean

    # --- Project Schemas ---
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        health:
          type: integer
        updatedAt:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    Segment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        constraints:
          type: array
          items:
            $ref: "#/components/schemas/Constraint"
        usedInFeatures:
          type: integer
        usedInProjects:
          type: integer
    CreateSegmentRequest:
      type: object
      required:
        - name
        - constraints
      properties:
        name:
          type: string
        description:
          type: string
        project:
          type: string
          description: Optional project scope
        constraints:
          type: array
          items:
            $ref: "#/components/schemas/Constraint"
    UpdateSegmentRequest:
      type: object
      properties:
        name:
          type: string
        constraints:
          type: array
          items:
            $ref: "#/components/schemas/Constraint"

    # --- Tag Schemas ---
...